Initial Commit.
Just a simple README and a submodule for requests. I see no reason not to keep
it close. I'm unlikely to do an official python package or bother with a
setup.py file before the summer, so it's better to let someone clone the
repository and have that there then to have to go looking for it on GitHub or in
PyPi.