Add issue functionality. Add github functionality
Add the ability to create an issue on a repository from the repository
view and from the GitHub view. Add ability to close and reopen an issue.

Added TODO file. Ironically forgot to add the following TODO to it.
Added an example to the README as well.

TODO: Finish issues()