Split Repository into it's own file.
Stop creating a Repository object in github3.issue for an Issue. Simply provide
the (login, repo_name) tuple so someone can fetch the Repository on their own.
Otherwise, issue.py will be a horribly large file with more than one section
inside of it. Do they have functionality which over laps? Certainly, but
following the API documentation, this is a saner way of organizing the module.