Finish Contents and start Downloads.
Added archive functionality with the option to specify a path to write the
archive to. Also added ability to get the readme and other file contents.

In models.py it was necessary to add a different get request handler so that
the headers and regular content would be accessible for the archive
functionality. I'm really going to need to refactor that area when I'm done
with everything else.

Also, now that my github account has my real name on it, I might as well start
using it in the README and setup.py as well.