Finish Watching. Start Hooks. Update TODO
I also realized that with some of the functions that check a condition, e.g.,
is_watching() and is_following(), GitHub returns 404 in the instance you're
not watching/following the repository/user. This would throw an Error instead
of returning a boolean value. Instead of catching the error, the simplest
solution is to simply check the status code returned by the request.