Events work. Missing property on Gist objects.
I realized I didn't have a property defined on the Gist objects (id property)
which is rather important. I realized this in the process of working on the
event types while working on Events.

Also, I modified setup.py to show the documentation as the homepage url and
changed classifies to a list as it is supposed to be.