Bring Gist object into line with module style.
On other objects, attributes like 'comments' usually indicate the number of
comments on that object where as methods like 'list_comments' actually return
the comments. Here, 'files' was an attribute returning a list of GistFiles
instead of the number of files. So I thought I should change that attribute to
the number of files on the Gist and add list_files() to return the list of
files.