Rename modules to prevent pollution of namespace.
Part of the problem with testing this has been that in the API there are
functions like user() or gist() which prevent me from being able to import
from github3.gist or github3.user the class(es) I need to make sure the right
objects are being returned. Changing the module names allows for no ambiguity
and (ideally) no clashes in names. This also means I can clean up __init__.py
which was becoming awfully cluttered.