Restrict API a bit. Expand tests.
I changed the organization of the repos tests. I may still further split up
the tests_methods method so that one of those won't cause the rest to be
ignored.

I realized why repo.hook() wasn't working last night. You need to be the
owner/a contributer for the repository for it to be visable to the user and
hence it requires authentication as do a few other related functions.

And I added a function to make testing class over a list easy.