Added coverage. Made tests loads faster.
Also I made the tests more unit like everywhere except test_issues. I need to
work on that after I finish testing github3.repos.

I made the tests a lot faster by defining my own __init__ functions on the
test classes. Telling the test class to get the repo in setUp causes it to get
it EVERY TIME, as opposed to just once. With that in mind, making it run once
in the __init__ makes the traffic go down on my network and makes the tests
faster and allows for more API calls in the future for tests.