Update TODO with my progress. Update .travis.yml
I'm part of the way to speeding up the test script. The time hog was in
loading the tests and I have somewhat made that parallel. It's just having
issues like in Travis and in py3k.

Also, Travis now supports excluding certain branches, and uricore will never
work but I may as well keep it on GitHub/BitBucket/etc. and Travis seems to
have tested it one too many times yesterday, so I just want to make sure it
doesn't happen again.

[ci skip]