Fixes #106
We were trying to send the list of Label objects to requests and json.dumps
was throwing a TypeError since the repr of a Label is <Label [labelname]>
instead of labelname. So we need to "normalize" the labels before sending them
to the JSON serializer.

Thanks to @alejandrogomez for the bug report. This is a huge issue.