Start work on validating parameters
This will give us greater flexibility with updating parameters on methods. For
example, I need to go through and find the methods where since should be
accepted. Instead of tagging it on to the end of the method signature (which I
despise), I can just add it the schema for the validator and let the **kwargs
and the validator do all the work. A class for the validator is probably
extreme and could probably be refactored to a function, but I'm not entirely
certain yet.