Add test for proper IssueEvent.issue
When the issue event JSON is obtained via the "List events for an issue"
API, it does not contain information about the parent issue.  In this
case the second argument to the IssueEvent constructor must be the
parent Issue, but Issue.iter_events() does not arrange for this.  As a
result IssueEvent.issue is bogus and __repr__() breaks.  Add a test for
this case.