Add Release.assets attribute
The JSON describing a release includes complete information on its
assets.  Add Release.assets attribute with a list of Assets.  This is
cheaper for the application than making a separate request via
Release.iter_assets().

Leave Release.iter_assets() in place, in case someone has a use for it,
but mention Release.assets in its documentation.