Stop moving around json data where response data is needed
To be able to store the response in created objects, and to remove the
ETag/Last-Modified kludges, let's ship response objects around and
update __init__, _json, and _instance_or_null to deal with this.

The GitHubObject's __init__ function can now handle both response
objects and json data, so code that wants to create more than one object
based on some JSON data still works.

Note that this is currently untested and only a few callsites of
_instance_or_null have been modified to demonstrate what I'm going for.