Support to assign multiple assignees to issue
Resolves https://github.com/sigmavirus24/github3.py/issues/626

I updated repo.py and issue.py to permit multiple assignees when
creating and editing an issue.  I modified _update_attributes, in the
issue class, to support the assignees attribute; I suspect, sometime in
the future, for GitHub to remove the attribute assignee alltogether.
Until then, we will continue to maintain the assignee attribute.

Although the GitHub's API supports the parameter assignee, it is
deprecated according to the documentation. See
https://developer.github.com/v3/issues/#create-an-issue for additional
details. Furthermore, we will continue to maintain the assignee
parameter, for backwards compatibility.

I added the standard unit and integration tests to support the changes.
In addition, I added two particular tests: assignee and assignees in the
request. GitHub's API responds with an exception: UnprocessableEntity.