Excise NullObject from github3.py
As part of on going work to improve the overall design and simplicity,
we're removing the NullObject and it's usage. At the present time, it
causes a lot of confusion for users (not as much as Empty, but still up
there) and there are simpler ways of designing the library than this.

Related-to #670