Add project column method to GitHub class
The Projects API allows a user to fetch a project column directly,
without reference to an organisation or repository. This commit
exposes this ability by adding a `project_column()` method to the
`GitHub` class.

Callers may obtain any project column to which they have access by
calling this method directly:

    import github3
    column = github3.login(token='token').project_column(1234)