Added Collaborator, Contributor and Repository.collaborators(affiliation).
See #730 for some background information.

Note that I have currently put the named parameter `affiliation` at the front
of the arguments in `Repository.collaborators`. I did this to be in line with
the other similar iterators. This can be moved to the back, but I assume some
backwards-incompatible changes are fine for 1.0.0?

This uses separate classes for the specific User object uses as I understood
the idea was from #670.