Split up Organization objects
This creates 3 new objects, one for iterations (Short), one for direct
GETs (Organization), and one for events (EventOrganization). Most
attributes can now be directly assigned, aside from a few that are only
returned in the API if they are set on the server. Default those to
None.

Many test cassettes needed to be updated to pick up attributes added to
orgs since the cassette was recorded (yay for additive REST APIs...).

Introduce a 'auto_login' method to handle either token auth or
username/password auth. Use it in the places I needed to update the
cassette.

Fix a few POST calls that were sending malformed (according to GitHub
API) json data. Specifically an empty permission is not accepted.

Related-to: #670

Signed-off-by: Jesse Keating <jkeating@j2solutions.net>