Split up PullRequest objects
This creates 3 classes; one for iterations (Short), one for direct GETs
(PullRequest), and one for events (EventPullRequest). Most attributes
are directly assigned, except where otherwise commented.

A couple cassettes needed to be updated, and the sample json for
pull_request needed to be updated as well. Updating that one required
updating the tests for the now current data.

Related-to #670

Signed-off-by: Jesse Keating <jkeating@j2solutions.net>