Add integration test for create_issue w/ assignees
There was not an integration test for this code path. Testing this path
uncovered a bug in ShortUser when attempting to repr such objects. The
name attribute is not available in the short object, so that object
should print differently.

The argument order for create_issue in github.py needed to match that of
create_issue in repo.py, since it just passes through to that function.

Signed-off-by: Jesse Keating <jkeating@j2solutions.net>