Clean up Release classes
This started as an effort to break up Release classes for "Short"
versions. However I discovered that there is no difference in data when
iterating over a Release or an Asset vs a direct GET of them. I still
cleaned up the classes to be more in line with other efforts.

Remove custom header as Releases are out of prerelease.

Update the example data sets with real release data from our own
repository.

Account for upload url differences from real data.

Related: #670

Signed-off-by: Jesse Keating <jkeating@j2solutions.net>