Raise exceptions for unexpected status codes
Also raise exceptions on any 400 or greater status code as discussed in
gh-670

As it stands today, it's hard for people to determine whether None is
returned because that's the API's response or because None is what we
return in the event of a 404 or some other "error"-like status code.

This change undoes that ambiguity and allows people to handle missing or
permission-related problems using exceptions.