Provide TokenAuth class to requests to ensure explicit usage
Provide a class based on BaseAuth that requests can all to set the
required headers on any calls made to ensure that it only uses the
given auth scheme.

In the event that `auth` is set to None, requests will default to
reading .netrc/_netrc files for any credentials set by the user. When
given an auth token or password it's necessary to ensure that it will
ignore any .netrc settings.