Start fleshing out updated branch protection API
This should pave the way for work in gh-696.

This adds a way to retrieve the information about a branch's protection.
It also begins to flesh out objects used to represent and update each
individual portions of a branch's protection.

To start with, there's support for managing required status checks and
the contexts that define them.