Add classes to support Checks API
This adds a new 'checks' module with classes to support the Checks API
(https://developer.github.com/v3/checks/). The CheckSuite and CheckRun
classes will be used to represent those objects.

There are some TODOs left in. Additional classes are needed to handle a
CheckRun output, which itelf needs classes for CheckRunAnnotation and
CheckRunAction. For now those are left in raw hash form. Additionally
when updating a CheckRun deeper cleaning of the parameters is needed, as
the parameters can include a list (of hashes) and a hash.

Testing has not been introduced yet. The APIs these classes use can only
be accessed by properly authenticated and authorized GitHub Apps
(https://developer.github.com/apps/about-apps/). When support for these
Apps lands I'm sure there will be new decorators to use to indicate the
requirement.

See #887 for further details.