Set timeout on HTTP requests
The requests documentation[0] recommends always setting a timeout
for requests in order to avoid indefinite hangs.  This can happen
if the remote side is unresponsive, or if a network error causes
a FIN or RST packet never to arrive.

This sets a default connect timeout of 4 seconds and a read timeout
of 1 second for all HTTP requests.  The default can be overridden in
the GitHubSession constructor.

The GitHubSession class is also added to the documentation on the
GitHub Object page since it is closely associated with it.

[0] http://docs.python-requests.org/en/master/user/quickstart/#timeouts