Add methods to retrieve issues and PRs from cards
While working on the projects and cards the other day, I realized that
trying to find the underlying issue or pull request was a nightmare. It
involved parsing the content url and then using a method on another
github3.py object to retrieve the object I wanted.

This creates methods that makes that work easier.