Increase default read timeout to 10 seconds
The read timeout is the time until Github starts the response to an
http request. The current default of one second can be too short for a
local maybe slower Github Enterprise. Instead change it to 10s as
github suggests also in its best practices for request from Github to
integrations [1].

[1] https://developer.github.com/v3/guides/best-practices-for-integrators/#favor-asynchronous-work-over-synchronous