Add support for Draft PR status in responses
This adds response data support for the `draft` status of a pull
request. This was recently
[added](https://github.blog/2019-02-14-introducing-draft-pull-requests/)
to the platform and API. A PullRequest can be marked as a `draft` which
will prevent merging and delay triggering automatic pull request reviews
from `CODEOWNERS`.

Support for receiving the data is still in preview so new headers have
to be littered throughout the codebase.

Resolves: #926