Avoid using content from NetRC when using GitHub App API
When using the default authentication handler and not overwriting
the auth parameter, the netrc file will be read and will overwrite
an existing value that was provided using the headers parameter.
As the value in the NetRC is wrong a 401 will happen. Using a
custom auth handler will ensure the expected token is beeing used.