Add support for head_repo param for PR creation
This parameter allows the repository containing the PR branch to
be fully specified in the head_repo with value of
`organization/repository` and using head param for the branch name.

This parameter is requried when the fork does not share the name as the
original repository. e.g. when the fork repositry is in the same
organization as the main repository.